Massive refactor to rush-eval. Moved all shell state into rush-state. Still need to add a crate for either builtins only or commands in general. Also still need to do a lot of post-refactor maintenance.
