+++ rush-eval/src/lib.rs
@@ -1,8 +1,3 @@
-#![allow(dead_code, unused_imports)]
-
 mod builtins;
-pub mod commands;
-mod environment;
-mod errors;
-mod path;
-pub mod shell;
+mod commands;
+pub mod evaluator;
