+++ rush-state/Cargo.toml
@@ -1,12 +1,10 @@
 [package]
-name = "rush-eval"
+name = "rush-state"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-rush-state = { path = "../rush-state" }
 anyhow.workspace = true
 thiserror.workspace = true
-colored.workspace = true
