+++ rush-state/src/config.rs
@@ -1,11 +1,8 @@
 use std::fs::File;
-use std::io::{stdin, stdout, BufRead, BufReader, Write};
+use std::io::{BufReader, BufRead};
 
 use anyhow::Result;
-use colored::Colorize;
 
-use crate::commands::{Context, Dispatcher};
-use crate::environment::Environment;
 use crate::errors::ShellError;
 
 // Represents any settings for the shell, most of which can be configured by the user
@@ -32,7 +29,7 @@ impl Default for Configuration {
 }
 
 impl Configuration {
-    fn from_file(filename: &str) -> Result<Self> {
+    pub fn from_file(filename: &str) -> Result<Self> {
         let mut config = Self::default();
         let file = File::open(filename).map_err(|_| ShellError::FailedToOpenConfigFile)?;
         let reader = BufReader::new(file);
@@ -78,79 +75,4 @@ impl Configuration {
 
         Ok(config)
     }
-}
-
-// Represents the shell, its state, and provides methods for interacting with it
-pub struct Shell {
-    pub environment: Environment,
-    pub config: Configuration,
-    dispatcher: Dispatcher,
-    success: bool,
-}
-
-impl Shell {
-    pub fn new() -> Result<Self> {
-        Ok(Self {
-            environment: Environment::new()?,
-            config: Configuration::from_file("config/config.rush")?,
-            dispatcher: Dispatcher::default(),
-            success: true,
-        })
-    }
-
-    // Evaluates and executes a command from a string
-    // $ Somewhat temporary, probably will be combined with .interpret()
-    pub fn eval(&mut self, command_name: String, command_args: Vec<String>) -> Result<()> {
-        self.interpret(command_name, command_args);
-        Ok(())
-    }
-
-    // Interprets a command from a string
-    fn interpret(&mut self, command_name: String, command_args: Vec<String>) {
-        let command_name = command_name.as_str();
-        let command_args = command_args.iter().map(|a| a.as_str()).collect();
-
-        // Bundle all the information that needs to be modifiable by the commands into a Context
-        let mut context = Context::new(&mut self.environment, &mut self.config);
-
-        // Dispatch the command to the Dispatcher
-        let exit_code = self.dispatcher.dispatch(command_name, command_args, &mut context);
-
-        // If the command was not found, print an error message
-        match exit_code {
-            Some(code) => {
-                self.success = code.is_ok();
-                if let Err(e) = code {
-                    if self.config.show_errors {
-                        eprintln!("Error: {}", format!("{:#?}: {}", e, e).red());
-                    }
-                }
-            }
-            None => {
-                eprintln!("Unknown command: {}", command_name.red());
-                self.success = false;
-            }
-        }
-    }
-}
-
-// Flushes stdout
-fn flush() -> Result<()> {
-    let mut stdout = stdout();
-    match stdout.flush() {
-        Ok(_) => Ok(()),
-        Err(_) => Err(ShellError::FailedToFlushStdout.into()),
-    }
-}
-
-// Reads a line of input from stdin
-fn read_line() -> Result<String> {
-    let mut line = String::new();
-    let stdin = stdin();
-    match stdin.read_line(&mut line) {
-        Ok(_) => (),
-        Err(_) => return Err(ShellError::FailedToReadStdin.into()),
-    }
-
-    Ok(line)
 }
